Routing continues, switched to 4-layer
Realized I needed the improved clearance/track widths of 4 layer service anyways for protos, and it was going to be a stretch routing this on a 2-layer design anyways. Switched over to 4 layer which makes life easier and should let me (with some stretching of DRC rules) break out the BGA codec chip.